- Interessenten und Bewerber in den Lehrverbandstree integriert - Senatsbeschlüsse sind jetzt in der Tabelle News - Jahresplan aktualisiert

+++ content/fasoverlay.js.php
@@ -190,46 +190,55 @@ function onVerbandSelect(event)
 	var grp=tree.view.getCellText(tree.currentIndex,col);
 	col = tree.columns ? tree.columns["gruppe"] : "gruppe";
 	var gruppe=tree.view.getCellText(tree.currentIndex,col);
-	
+	col = tree.columns ? tree.columns["typ"] : "typ";
+	var typ=tree.view.getCellText(tree.currentIndex,col);
+	col = tree.columns ? tree.columns["stsem"] : "stsem";
+	var stsem=tree.view.getCellText(tree.currentIndex,col);
+		
 	currentAuswahl.stg_kz=stg_kz;
 	currentAuswahl.sem=sem;
 	currentAuswahl.ver=ver;
 	currentAuswahl.grp=grp;
 	currentAuswahl.gruppe=gruppe;
 
-	// Studenten
-	try
+	if(typ=='')
 	{
-		url = "<?php echo APP_ROOT; ?>rdf/student.rdf.php?"+"stg_kz="+stg_kz+"&sem="+sem+"&ver="+ver+"&grp="+grp+"&gruppe="+gruppe+"&stsem=true&"+gettimestamp();
-		var treeStudent=document.getElementById('student-tree');
-		
-		//Alte DS entfernen
-		var oldDatasources = treeStudent.database.GetDataSources();
-		while(oldDatasources.hasMoreElements())
+		// Studenten
+		try
 		{
-			treeStudent.database.RemoveDataSource(oldDatasources.getNext());
+			url = "<?php echo APP_ROOT; ?>rdf/student.rdf.php?"+"stg_kz="+stg_kz+"&sem="+sem+"&ver="+ver+"&grp="+grp+"&gruppe="+gruppe+"&stsem=true&"+gettimestamp();
+			var treeStudent=document.getElementById('student-tree');
+			
+			//Alte DS entfernen
+			var oldDatasources = treeStudent.database.GetDataSources();
+			while(oldDatasources.hasMoreElements())
+			{
+				treeStudent.database.RemoveDataSource(oldDatasources.getNext());
+			}
+	
+			var rdfService = Components.classes["@mozilla.org/rdf/rdf-service;1"].getService(Components.interfaces.nsIRDFService);
+			StudentTreeDatasource = rdfService.GetDataSource(url);
+			StudentTreeDatasource.QueryInterface(Components.interfaces.nsIRDFRemoteDataSource);
+			StudentTreeDatasource.QueryInterface(Components.interfaces.nsIRDFXMLSink);
+			treeStudent.database.AddDataSource(StudentTreeDatasource);
+			StudentTreeDatasource.addXMLSinkObserver(StudentTreeSinkObserver);
+			treeStudent.builder.addListener(StudentTreeListener);
+			StudentDetailReset();
+			StudentDetailDisableFields(true);
+			StudentPrestudentDisableFields(true);
 		}
-
-		var rdfService = Components.classes["@mozilla.org/rdf/rdf-service;1"].getService(Components.interfaces.nsIRDFService);
-		StudentTreeDatasource = rdfService.GetDataSource(url);
-		StudentTreeDatasource.QueryInterface(Components.interfaces.nsIRDFRemoteDataSource);
-		StudentTreeDatasource.QueryInterface(Components.interfaces.nsIRDFXMLSink);
-		treeStudent.database.AddDataSource(StudentTreeDatasource);
-		StudentTreeDatasource.addXMLSinkObserver(StudentTreeSinkObserver);
-		treeStudent.builder.addListener(StudentTreeListener);
-		StudentDetailReset();
-		StudentDetailDisableFields(true);
-		StudentPrestudentDisableFields(true);
-	}
-	catch(e)
-	{
-		debug(e);
+		catch(e)
+		{
+			debug(e);
+		}
 	}
 	
 	// Interessenten / Bewerber
 	try
 	{
-		url = "<?php echo APP_ROOT; ?>rdf/interessentenbewerber.rdf.php?"+"studiengang_kz="+stg_kz+"&semester="+sem+"&"+gettimestamp();
+		if(stsem=='' && typ=='')
+			stsem='aktuelles';
+		url = "<?php echo APP_ROOT; ?>rdf/interessentenbewerber.rdf.php?"+"studiengang_kz="+stg_kz+"&semester="+sem+"&typ="+typ+"&studiensemester_kurzbz="+stsem+"&"+gettimestamp();
 		var treeInt=document.getElementById('interessent-tree');
 		
 		//Alte DS entfernen
@@ -255,32 +264,35 @@ function onVerbandSelect(event)
 		debug(e);
 	}
 
-	// Lehrveranstaltung
-	try
+	if(typ=='')
 	{
-		url = '<?php echo APP_ROOT; ?>rdf/lehrveranstaltung_einheiten.rdf.php?stg_kz='+stg_kz+'&sem='+sem+'&ver='+ver+'&grp='+grp+'&gruppe='+gruppe+'&'+gettimestamp();
-		var treeLV=document.getElementById('lehrveranstaltung-tree');
-
-		//Alte DS entfernen
-		var oldDatasources = treeLV.database.GetDataSources();
-		while(oldDatasources.hasMoreElements())
+		// Lehrveranstaltung
+		try
 		{
-			treeLV.database.RemoveDataSource(oldDatasources.getNext());
+			url = '<?php echo APP_ROOT; ?>rdf/lehrveranstaltung_einheiten.rdf.php?stg_kz='+stg_kz+'&sem='+sem+'&ver='+ver+'&grp='+grp+'&gruppe='+gruppe+'&'+gettimestamp();
+			var treeLV=document.getElementById('lehrveranstaltung-tree');
+	
+			//Alte DS entfernen
+			var oldDatasources = treeLV.database.GetDataSources();
+			while(oldDatasources.hasMoreElements())
+			{
+				treeLV.database.RemoveDataSource(oldDatasources.getNext());
+			}
+	
+			var rdfService1 = Components.classes["@mozilla.org/rdf/rdf-service;1"].getService(Components.interfaces.nsIRDFService);
+	
+			LvTreeDatasource = rdfService1.GetDataSource(url);
+			LvTreeDatasource.QueryInterface(Components.interfaces.nsIRDFRemoteDataSource);
+			LvTreeDatasource.QueryInterface(Components.interfaces.nsIRDFXMLSink);
+			treeLV.database.AddDataSource(LvTreeDatasource);
+			LvTreeDatasource.addXMLSinkObserver(LvTreeSinkObserver);
+			treeLV.builder.addListener(LvTreeListener);
+			document.getElementById('lehrveranstaltung-toolbar-lehrauftrag').hidden=true;
 		}
-
-		var rdfService1 = Components.classes["@mozilla.org/rdf/rdf-service;1"].getService(Components.interfaces.nsIRDFService);
-
-		LvTreeDatasource = rdfService1.GetDataSource(url);
-		LvTreeDatasource.QueryInterface(Components.interfaces.nsIRDFRemoteDataSource);
-		LvTreeDatasource.QueryInterface(Components.interfaces.nsIRDFXMLSink);
-		treeLV.database.AddDataSource(LvTreeDatasource);
-		LvTreeDatasource.addXMLSinkObserver(LvTreeSinkObserver);
-		treeLV.builder.addListener(LvTreeListener);
-		document.getElementById('lehrveranstaltung-toolbar-lehrauftrag').hidden=true;
-	}
-	catch(e)
-	{
-		debug(e);
+		catch(e)
+		{
+			debug(e);
+		}
 	}
 }
